+++ service/queue/schema/360$queue.queue.sql
@@ -1,7 +1,7 @@
 CREATE TABLE queue."queue"  (
     "queueId" BIGSERIAL,
     "recordId" BIGINT,
-    "retryId" INTEGER,
+    "retryId" SMALLINT,
     "expirationDate" timestamp without time zone,
     "updatedAt" timestamp without time zone NOT NULL,
     CONSTRAINT "pkQueueQueue" PRIMARY KEY ("queueId"),
